Match Power BI connector to existing architecture
